🚚 api: move route to src/team context
