+++ api/lib/application/organizations/index.js
@@ -141,46 +141,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'POST',
-      path: '/api/admin/organizations/{id}/invitations',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: organizationController.sendInvitationByLangAndRole,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-          options: {
-            allowUnknown: true,
-          },
-          payload: Joi.object({
-            data: {
-              attributes: {
-                email: Joi.string().email().required(),
-                lang: Joi.string().valid('fr-fr', 'fr', 'en'),
-                role: Joi.string().valid('ADMIN', 'MEMBER').allow(null),
-              },
-            },
-          }),
-        },
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            "- Elle permet d'inviter des personnes, déjà utilisateurs de Pix ou non, à être membre d'une organisation, via leur **email**",
-        ],
-        tags: ['api', 'invitations'],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/admin/organizations/{id}/target-profile-summaries',
       config: {
